Add more comments and change file name of fs related type conversion
