+++ src/typemap/src/lib.rs
@@ -1,8 +1,7 @@
-pub mod path_conv;
+pub mod fs_conv;
 pub mod syscall_conv;
 pub mod type_conv;
 
-pub use path_conv::*;
+pub use fs_conv::*;
 pub use syscall_conv::*;
 pub use type_conv::*;
-pub use type_conv::get_pipearray;
