+++ src/typemap/src/syscall_conv.rs
@@ -1,8 +1,9 @@
-//! Top level Type Conversion API
+//! Basic primitive type conversion API
 //!
-//! This file provides the top level type conversion API needed for actual syscall implementation
-//! under src/syscalls/
-use crate::path_conv::*;
+//! This file defines conversion helpers for basic primitive types (e.g., `i32`, `u32`, `i64`).
+//! These functions are used during syscall argument decoding and type-safe interpretation
+//! within the RawPOSIX syscall layer (`src/syscalls/`).
+use crate::fs_conv::*;
 use crate::type_conv::*;
 use cage::get_cage;
 use cage::memory::mem_helper::*;
@@ -12,28 +13,6 @@ use std::str::Utf8Error;
 use sysdefs::constants::err_const::{syscall_error, Errno};
 use sysdefs::constants::fs_const::{MAX_CAGEID, PATH_MAX};
 
-/// Translate a received virtual file descriptor (`virtual_fd`) to real kernel file descriptor.
-/// This function is not for security purpose. Always using arg_cageid to translate.
-///     - If arg_cageid != cageid: this call is sent by grate. We need to translate according to cage
-///     - If arg_cageid == cageid: this call is sent by cage, we can use either one
-/// Return: underlying kernel file descriptor
-pub fn convert_fd_to_host(virtual_fd: u64, arg_cageid: u64, cageid: u64) -> i32 {
-    #[cfg(feature = "secure")]
-    {
-        if !validate_cageid(path_arg_cageid, cageid) {
-            return -1;
-        }
-    }
-    // Find corresponding virtual fd instance from `fdtable` subsystem
-    let wrappedvfd = fdtables::translate_virtual_fd(arg_cageid, virtual_fd);
-    if wrappedvfd.is_err() {
-        return -9;
-    }
-    let vfd = wrappedvfd.unwrap();
-    // Actual kernel fd mapped with provided virtual fd
-    vfd.underfd as i32
-}
-
 /// This function provides two operations: first, it translates path pointer address from WASM environment
 /// to kernel system address; then, it adjusts the path from user's perspective to host's perspective,
 /// which is adding `LIND_ROOT` before the path arguments. Considering actual syscall implementation
@@ -42,13 +21,13 @@ pub fn convert_fd_to_host(virtual_fd: u64, arg_cageid: u64, cageid: u64) -> i32 
 /// implementation)
 ///     - If arg_cageid != cageid: this call is sent by grate. We need to translate according to cage
 ///     - If arg_cageid == cageid: this call is sent by cage, we can use either one
-/// 
-/// ##Arguments:
+///
+/// ## Arguments:
 ///     - cageid: required to do address translation for path pointer
 ///     - path_arg: the path pointer with address and contents from user's perspective. Address is
 ///                 32-bit (because of WASM feature).
 ///
-/// ##Outputs:
+/// ## Returns:
 ///     - c_path: a `CString` variable stores the path from host's perspective
 ///     - will return error if total length exceed the MAX_PATH (which is 4096). We use `Box<dyn Error>` here to
 ///      let upper functions do error handling. (ie: we want to )
@@ -68,72 +47,40 @@ pub fn get_cstr<'a>(arg: u64) -> Result<&'a str, i32> {
 
     return Err(-1);
 }
-pub fn sc_convert_path_to_host(path_arg: u64, path_arg_cageid: u64, cageid: u64) -> CString {
-    #[cfg(feature = "secure")]
-    {
-        if !validate_cageid(path_arg_cageid, cageid) {
-            panic!("Invalide Cage ID");
-        }
-    }
-    let cage = get_cage(path_arg_cageid).unwrap();
-    let addr = translate_vmmap_addr(&cage, path_arg).unwrap();
-    let path = match get_cstr(addr) {
-        Ok(path) => path,
-        Err(e) => panic!("{:?}", e),
-    };
-    // We will create a new variable in host process to handle the path value
-    let relpath = normpath(convpath(path), path_arg_cageid);
-    let relative_path = relpath.to_str().unwrap();
-
-    #[cfg(feature = "secure")]
-    {
-        let total_length = LIND_ROOT.len() + relative_path.len();
-
-        if total_length >= PATH_MAX {
-            panic!("Path exceeds PATH_MAX (4096)");
-        }
-    }
-
-    // CString will handle the case when string is not terminated by `\0`, but will return error if `\0` is
-    // contained within the string.
-    let full_path = format!("{}{}", LIND_ROOT, relative_path);
-    match CString::new(full_path) {
-        Ok(c_path) => c_path,
-        Err(_) => panic!("String contains internal null byte"),
-    }
-}
 
 /// Validate whether two cage ids are in valid range. This is used for security mode in
 /// type conversion.
-/// 
+///
 /// ## Arguments:
 /// cageid_1: first cage id
 /// cageid_2: second cage id
-/// 
-/// ## Output:
+///
+/// ## Returns:
 /// true: both of them are valid
-/// false: one of them or neither of them are valid 
+/// false: one of them or neither of them are valid
 pub fn validate_cageid(cageid_1: u64, cageid_2: u64) -> bool {
-    if cageid_1 > MAX_CAGEID as u64 || cageid_2 > MAX_CAGEID as u64 {
+    if cageid_1 > MAX_CAGEID as u64 || cageid_2 > MAX_CAGEID as u64 || cageid_1 < 0 || cageid_2 < 0
+    {
         return false;
     }
     true
 }
 
-/// `sc_convert_sysarg_to_i32` is the type conversion function used to convert the 
-/// argument's type from u64 to i32. When in `secure` mode, extra checks will be 
-/// performed through `get_i32()` function. (for example validating if all upper-bit 
+/// `sc_convert_sysarg_to_i32` is the type conversion function used to convert the
+/// argument's type from u64 to i32. When in `secure` mode, extra checks will be
+/// performed through `get_i32()` function. (for example validating if all upper-bit
 /// are 0; if cage ids are in valid range). The security mode can be enabled through
-/// compilation flag of this library. Those calls will panic when failed the check 
+/// compilation flag of this library. Those calls will panic when failed the check
 /// for security concerns  
-/// 
+///
 /// `get_i32()`
+///
 /// ## Arguments:
 /// arg: the argument value
 /// arg_cageid: the argument's cageid
 /// cageid: source cageid (the cage execute this call)
-/// 
-/// ## Output:
+///
+/// ## Returns:
 /// Success: A converted i32
 /// Fail: panic
 pub fn get_i32(arg: u64, arg_cageid: u64, cageid: u64) -> i32 {
@@ -152,8 +99,8 @@ pub fn get_i32(arg: u64, arg_cageid: u64, cageid: u64) -> i32 {
 /// arg: the argument value
 /// arg_cageid: the argument's cageid
 /// cageid: source cageid (the cage execute this call)
-/// 
-/// ## Output:
+///
+/// ## Returns:
 /// Success: A converted i32
 /// Fail: panic
 pub fn sc_convert_sysarg_to_i32(arg: u64, arg_cageid: u64, cageid: u64) -> i32 {
@@ -164,20 +111,20 @@ pub fn sc_convert_sysarg_to_i32(arg: u64, arg_cageid: u64, cageid: u64) -> i32 {
     return get_i32(arg, arg_cageid, cageid);
 }
 
-/// `sc_convert_sysarg_to_u32` is the type conversion function used to convert the 
-/// argument's type from u64 to u32. When in `secure` mode, extra checks will be 
-/// performed through `get_u32()` function. (for example validating if all upper-bit 
+/// `sc_convert_sysarg_to_u32` is the type conversion function used to convert the
+/// argument's type from u64 to u32. When in `secure` mode, extra checks will be
+/// performed through `get_u32()` function. (for example validating if all upper-bit
 /// are 0; if cage ids are in valid range). The security mode can be enabled through
-/// compilation flag of this library. Those calls will panic when failed the check 
+/// compilation flag of this library. Those calls will panic when failed the check
 /// for security concerns  
-/// 
+///
 /// `get_u32()`
 /// ## Arguments:
 /// arg: the argument value
 /// arg_cageid: the argument's cageid
 /// cageid: source cageid (the cage execute this call)
-/// 
-/// ## Output:
+///
+/// ## Returns:
 /// Success: A converted u32
 /// Fail: panic
 pub fn get_u32(arg: u64, arg_cageid: u64, cageid: u64) -> u32 {
@@ -196,8 +143,8 @@ pub fn get_u32(arg: u64, arg_cageid: u64, cageid: u64) -> u32 {
 /// arg: the argument value
 /// arg_cageid: the argument's cageid
 /// cageid: source cageid (the cage execute this call)
-/// 
-/// ## Output:
+///
+/// ## Returns:
 /// Success: A converted u32
 /// Fail: panic
 pub fn sc_convert_sysarg_to_u32(arg: u64, arg_cageid: u64, cageid: u64) -> u32 {
@@ -208,18 +155,18 @@ pub fn sc_convert_sysarg_to_u32(arg: u64, arg_cageid: u64, cageid: u64) -> u32 {
     return get_u32(arg);
 }
 
-/// `sc_convert_sysarg_to_isize` is the type conversion function used to convert the 
-/// argument's type from u64 to isize. When in `secure` mode, extra checks will be 
-/// performed through `validate_cageid()` function (whether cage ids are in valid 
-/// range). The security mode can be enabled through compilation flag of this library. 
+/// `sc_convert_sysarg_to_isize` is the type conversion function used to convert the
+/// argument's type from u64 to isize. When in `secure` mode, extra checks will be
+/// performed through `validate_cageid()` function (whether cage ids are in valid
+/// range). The security mode can be enabled through compilation flag of this library.
 /// Those calls will panic when failed the check for security concerns  
-/// 
+///
 /// ## Arguments:
 /// arg: argument value
 /// arg_cageid: argument's cageid
 /// cageid: source cageid (the cage executes the call)
-/// 
-/// ## Output: 
+///
+/// ## Returns:
 /// Success: A converted isize
 /// Fail: panic
 pub fn sc_convert_sysarg_to_isize(arg: u64, arg_cageid: u64, cageid: u64) -> isize {
@@ -232,18 +179,18 @@ pub fn sc_convert_sysarg_to_isize(arg: u64, arg_cageid: u64, cageid: u64) -> isi
     }
 }
 
-/// `sc_convert_sysarg_to_usize` is the type conversion function used to convert the 
-/// argument's type from u64 to usize. When in `secure` mode, extra checks will be 
-/// performed through `validate_cageid()` function (whether cage ids are in valid 
-/// range). The security mode can be enabled through compilation flag of this library. 
+/// `sc_convert_sysarg_to_usize` is the type conversion function used to convert the
+/// argument's type from u64 to usize. When in `secure` mode, extra checks will be
+/// performed through `validate_cageid()` function (whether cage ids are in valid
+/// range). The security mode can be enabled through compilation flag of this library.
 /// Those calls will panic when failed the check for security concerns  
-/// 
+///
 /// ## Arguments:
 /// arg: argument value
 /// arg_cageid: argument's cageid
 /// cageid: source cageid (the cage executes the call)
-/// 
-/// ## Output: 
+///
+/// ## Returns:
 /// Success: A converted usize
 /// Fail: panic
 pub fn sc_convert_sysarg_to_usize(arg: u64, arg_cageid: u64, cageid: u64) -> usize {
@@ -256,18 +203,18 @@ pub fn sc_convert_sysarg_to_usize(arg: u64, arg_cageid: u64, cageid: u64) -> usi
     }
 }
 
-/// `sc_convert_sysarg_to_i64` is the type conversion function used to convert the 
-/// argument's type from u64 to i64. When in `secure` mode, extra checks will be 
-/// performed through `validate_cageid()` function (whether cage ids are in valid 
-/// range). The security mode can be enabled through compilation flag of this library. 
+/// `sc_convert_sysarg_to_i64` is the type conversion function used to convert the
+/// argument's type from u64 to i64. When in `secure` mode, extra checks will be
+/// performed through `validate_cageid()` function (whether cage ids are in valid
+/// range). The security mode can be enabled through compilation flag of this library.
 /// Those calls will panic when failed the check for security concerns  
-/// 
+///
 /// ## Arguments:
 /// arg: argument value
 /// arg_cageid: argument's cageid
 /// cageid: source cageid (the cage executes the call)
-/// 
-/// ## Output: 
+///
+/// ## Returns:
 /// Success: A converted i64
 /// Fail: panic
 pub fn sc_convert_sysarg_to_i64(arg: u64, arg_cageid: u64, cageid: u64) -> i64 {
@@ -281,16 +228,16 @@ pub fn sc_convert_sysarg_to_i64(arg: u64, arg_cageid: u64, cageid: u64) -> i64 {
 }
 
 /// `sc_unusedarg()` is the security check function used to validate all unused args. This
-/// will return true in default mode, and check if `arg` with `arg_cageid` are all null in 
+/// will return true in default mode, and check if `arg` with `arg_cageid` are all null in
 /// `secure` mode.
-/// 
+///
 /// ## Arguments:
 /// arg: argument value
 /// arg_cageid: argument's cageid
-/// 
-/// ## Output:
+///
+/// ## Returns:
 /// Always true in default mode.
-/// In secure mode: 
+/// In secure mode:
 /// Return true if all null, false otherwise.
 pub fn sc_unusedarg(arg: u64, arg_cageid: u64) -> bool {
     #[cfg(feature = "fast")]
@@ -307,7 +254,7 @@ pub fn sc_unusedarg(arg: u64, arg_cageid: u64) -> bool {
 ///     - cageid: required to do address translation for buf pointer
 ///     - buf_arg: the buf pointer address, which is 32-bit because of WASM feature
 ///
-/// ## Output:
+/// ## Returns:
 ///     - buf: actual system address, which is the actual position that stores data
 pub fn sc_convert_buf(buf_arg: u64, arg_cageid: u64, cageid: u64) -> *const u8 {
     // Get cage reference to translate address
